web/requirements.txt
arrow==0.6.0
Django==1.8.4
django-filter==0.11.0
django-js-reverse==0.6.1
django-mptt==0.7.4
djangorestframework==3.2.3
flake8==2.4.1
Markdown==2.6.2
mccabe==0.3.1
MySQL-python==1.2.5
pep8==1.6.2
Pillow==3.0.0
pyflakes==0.9.2
python-dateutil==2.4.2
gunicorn==19.3.0
six==1.9.0
unipath==1.1

django-extensions==1.7.4
django_fagungis==0.1
psycopg2==2.6.1